Basic Menu Callback is in place.
We are now able to call member functions from other
member functions using a callbacks via function pointers.

This allows EditorMenu to notify its parent that it is done
with something. The parent will be able to handle what to
do next, the EditorMenu does not have to know anything
about TextEditor.

This starts work on issue #15

Signed-off-by: Isaac Travers <isaac.a.travers@gmail.com>