Create Member Function Callback System for Menu
When creating a menu item, it is now possible to register
a callback function can will be called if that menu
item is clicked.

We have implemented the ability to click on a
Horizontal Menu, but have not yet implemented that
ability to click on a vertical menu.

This finishes Issue #15
and continues work on issues #1 #5 #6 and #9

Signed-off-by: Isaac Travers <isaac.a.travers@gmail.com>