Items are now being added to menu bar semi-correctly
Everything is working except there is some sort of
bug in my submenu placement code, other than that
the submenus are now wired up to something that
can be extended. So this finishes issue #5 .

MenuBar.itemItem() bug is taken care of here.
So this also finishes issue #17

I will also say that this officially takes care of
issue #6 investigating UI components is good.

This commit partially addresses issues #1 and #16

Signed-off-by: Isaac Travers <isaac.a.travers@gmail.com>