Create an OpenHuffman Menu Option
A menu option that gives the user the ability to open a specific file.

We currently have a dialog box pop up that allows the user to choose the
filename.

The dialog box is buggy though. So I created issue #38 to fix it.
This commit partially addresses issue #28
and is work on Milestone 6

Signed-off-by: Isaac Travers <isaac.a.travers@gmail.com>