Started with code to encode our text
I started with the code to encode our text,
It looks like our huffman tree is encoding
with the prefix property, but it doesn't
look like I was expecting.

I was expecting the left child of any node to only be
a leaf node, whereas that is not happening. I'm not sure
if thats a bug, or if it needs more research.

THis partially addresses issue #33

Signed-off-by: Isaac A Travers <iat6@humboldt.edu>